chore(plugins/plugin-kubectl): Refactor to defer loading of xtermjs components
part of #8066
