+++ plugins/plugin-kubectl/src/lib/view/modes/logs-mode.tsx
@@ -14,19 +14,36 @@
  * limitations under the License.
  */
 
-import { i18n, Arguments, ModeDeclaration, ModeRegistration, Tab } from '@kui-shell/core'
+import React from 'react'
+import { i18n, Arguments, ModeDeclaration, ModeRegistration, Tab, ToolbarProps } from '@kui-shell/core'
 
-import { Pod, isPod, isDeployment, isReplicaSet } from '../../model/resource'
+import mode from './logs-mode-id'
 import { KubeOptions } from '../../../controller/kubectl/options'
+import { Pod, isPod, Deployment, isDeployment, ReplicaSet, isReplicaSet } from '../../model/resource'
+
+const Logs = React.lazy(() => import('./Logs'))
 
 const strings = i18n('plugin-kubectl', 'logs')
 
-export const mode: ModeDeclaration<Pod> = {
-  mode: 'logs',
+/**
+ * The content renderer for the summary tab
+ *
+ */
+export async function content(tab: Tab, resource: Pod | Deployment | ReplicaSet, args: Arguments<KubeOptions>) {
+  return {
+    react: function LogsProvider(toolbarController: ToolbarProps) {
+      return <Logs tab={tab} mode={mode} resource={resource} args={args} toolbarController={toolbarController} />
+    }
+  }
+}
+
+export const modeDecl: ModeDeclaration<Pod> = {
+  mode,
   label: strings('Logs'),
 
   // hack workaround to defer loading of plugin-client-common; needs update to babel-plugin-ignore-html-and-css-imports
-  content: (tab: Tab, pod: Pod, args: Arguments<KubeOptions>) => import('./logs').then(_ => _.content(tab, pod, args))
+  // content: (tab: Tab, pod: Pod, args: Arguments<KubeOptions>) => import('./logs').then(_ => _.content(tab, pod, args))
+  content
 }
 
 /**
@@ -36,7 +53,7 @@ export const mode: ModeDeclaration<Pod> = {
  */
 const logsReg: ModeRegistration<Pod> = {
   when: _ => isPod(_) || isDeployment(_) || isReplicaSet(_),
-  mode
+  mode: modeDecl
 }
 
 export default logsReg
